Made a lot of updates

+++ app/Livewire/EditMenuSection.php
@@ -8,6 +8,7 @@ use App\Models\MenuSections;
 use App\Models\MenuItems;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Str;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 
 class EditMenuSection extends Component
@@ -38,40 +39,38 @@ class EditMenuSection extends Component
         $this->menuItems[] = ['name' => '', 'price' => '', 'image'=> null];
     }
 
-    public function removeMenuItem($index, $itemID)
-    {
-        \Log::info('Attempting to remove menu item with section id: '.$itemID);
-    
-        try {
-        $item = MenuItems::findOrFail($itemID);
+
+    public function removeMenuItem($index, $itemID = null){
+
+        \Log::debug('executing function '.__FUNCTION__.'');
+        \Log::debug('checking if ItemID '.$itemID. ' exists in the DB...');
         
-        if ($item) {
-            \Log::info('Item found, checking if an image exists...');
-            if ($item->image_path && Storage::disk('public')->exists($item->image_path)) {
-                \Log::info('Image exists: '.$item->image_path);
-                \Log::info('Deleting image...');
-                Storage::disk('public')->delete($item->image_path);
-                \Log::info('Image deleted!');
+        if($itemID){
+            $item = MenuItems::find($itemID);
+            \Log::debug('ItemID found! Checking if item exists...');
+
+            if($item){
+
+                \Log::debug('Item exists! checking if item has an associated image in the DB and on filesystem...');
+
+                if($item->image_path && Storage::disk('public')->exists($item->image_path)){
+                    Storage::disk('public')->delete($item->image_path);
+                    \Log::debug('image found and deleted!');
+                }
+
+                $item->delete();
+                \Log::debug('Item deleted from DB!');
             }
-    
-            \Log::info('Deleting item...');
-            $item->delete();
-            \Log::info('Item deleted!');
+
+            else{
+                \Log::debug('no item found in db, deleting from array...');
+                // if no item found in DB, remove from array and re-index
+                unset($this->menuItems[$index]);
+                $this->menuItems = array_values($this->menuItems);
+                \Log::debug('item deleted from the array!');
+            }
         }
-    
-        \Log::info('Unsetting menu item index and re-indexing the array...');
-        unset($this->menuItems[$index]);
-        $this->menuItems = array_values($this->menuItems);
-        \Log::info('Item index removed and array re-indexed.');
-
     }
-    catch(ModelNotFoundException $e){
-        \Log::info('Item not found');
-        $this->error = 'Cannot find item to delete';
-    }
-
-    }
-    
 
     public function save(){
         $this->validate([
@@ -81,76 +80,55 @@ class EditMenuSection extends Component
             'menuItems.*.price' => 'required|numeric|min:0',
             'menuItems.*.image' => 'nullable|image|mimes:jpeg,png,jpg,gif|max:2048',
         ]);
-    
+
         try {
-            \Log::info('Editing Section: '.$this->section->name);
-    
+
             $this->section->name = $this->name;
             $this->section->status = $this->status;
-    
-            \Log::info('Data to save: '. $this->name. ' and '.$this->status);
-            
-            \Log::info('Checking to see if an image is an instance of the UploadedFile class...');
-            
-            if ($this->image instanceof UploadedFile) {
-                \Log::info('It is, now checking to see if we have an existing image...');
-                
-                if ($this->section->section_image && Storage::exists('public/' . $this->section->section_image)) {
-                    \Log::info('Found old image! Deleting old image...');
-                    Storage::delete('public/' . $this->section->section_image);
-                    \Log::info('Image deleted from the server!');
+
+            if($this->image instanceof UploadedFile){
+                if($this->section->section_image && Storage::exists('public/'.$this->section->section_image)){
+                    Storage::delete('public/'.$this->section->section_image);
+
                 }
-    
-                // Store new image
-                \Log::info('Saving image...');
+
                 $imagePath = $this->image->store('section_image', 'public');
                 $this->section->section_image = $imagePath;
-                \Log::info('Saved image location at: ' . $imagePath);
             }
-    
-            // Save the section first
+
             $this->section->save();
-            \Log::info('Data saved for menu section!');
-    
-            \Log::info('Checking to see if the user has menu items to save...');
-            if (isset($this->menuItems)) {
-                \Log::info('User does have menu items to save. Iterating over menu items...');
-    
-                foreach ($this->menuItems as $item) {
-                    \Log::info('Saving menu item data...');
-                    \Log::info('LOGGING MENU ITEM: ' . json_encode($item));
-    
-                    // Check if the menu item image exists and is an instance of UploadedFile
-                    if (isset($item['image']) && $item['image'] instanceof UploadedFile) {
-                        // If an image is provided, store it
+
+
+            if(isset($this->menuItems)){
+
+                foreach($this->menuItems as $item){
+                    if(isset($item['image']) && $item['image'] instanceof UploadedFile){
                         $imagePath = $item['image']->store('section_images', 'public');
-                    } else {
-                        // If no image is provided, do not change the existing image
+                    }
+
+                    else {
                         $imagePath = isset($item['image_path']) ? $item['image_path'] : null;
                     }
-    
-                    // Update or create the menu item
-                    $menuItem = MenuItems::updateOrCreate(
-                        [
-                            'menu_section_id'=>$this->section->menu_section_id,
-                            'name' => $item['name'],
-                            'price' => $item['price'],
-                            'image_path' => $imagePath,  // Only update image if provided
-                        ]
-                    );
+
+                    $menuItem = MenuItems::updateOrCreate([
+                        'itemID'=> $item['itemID'] ?? Str::uuid(5),
+                        'menu_section_id'=> $this->section->menu_section_id,
+                        'name'=>$item['name'],
+                        'price'=>$item['price'],
+                        'image_path'=>$imagePath,
+                    ]);
+
                 }
-    
-                \Log::info('Menu item data saved!');
             }
-    
             $this->success = 'Menu section updated successfully';
-    
-        } catch (\Exception $e) {
+        }
+        
+        catch(\Exception $e){
             \Log::error('Error saving menu section in class: ' . __CLASS__ . ' in method: ' . __FUNCTION__ . ' on line: ' . __LINE__ . ' error message: ' . $e->getMessage());
+
             $this->error = 'An error occurred while saving changes';
         }
     }
-    
 
     public function render()
     {
